Relocate transaction_manager
