+++ Makefile.am
@@ -53,11 +53,11 @@ cgimap_apidb_include_HEADERS = \
 	include/cgimap/backend/apidb/quad_tile.hpp \
 	include/cgimap/backend/apidb/readonly_pgsql_selection.hpp \
 	include/cgimap/backend/apidb/writeable_pgsql_selection.hpp \
+ 	include/cgimap/backend/apidb/transaction_manager.hpp \
 	include/cgimap/backend/apidb/utils.hpp \
 	include/cgimap/backend/apidb/changeset_upload/changeset_updater.hpp \
 	include/cgimap/backend/apidb/changeset_upload/node_updater.hpp \
 	include/cgimap/backend/apidb/changeset_upload/relation_updater.hpp \
- 	include/cgimap/backend/apidb/changeset_upload/transaction_manager.hpp \
 	include/cgimap/backend/apidb/changeset_upload/way_updater.hpp
 endif
