+++ include/cgimap/backend/apidb/changeset_upload/changeset_updater.hpp
@@ -6,7 +6,7 @@
 
 #include "cgimap/api06/changeset_upload/changeset_updater.hpp"
 #include "cgimap/api06/changeset_upload/osmchange_tracking.hpp"
-#include "cgimap/backend/apidb/changeset_upload/transaction_manager.hpp"
+#include "cgimap/backend/apidb/transaction_manager.hpp"
 
 class ApiDB_Changeset_Updater : public Changeset_Updater {
