+++ include/cgimap/backend/apidb/changeset_upload/relation_updater.hpp
@@ -7,7 +7,7 @@
 #include "cgimap/api06/changeset_upload/osmchange_tracking.hpp"
 #include "cgimap/api06/changeset_upload/relation.hpp"
 #include "cgimap/api06/changeset_upload/relation_updater.hpp"
-#include "cgimap/backend/apidb/changeset_upload/transaction_manager.hpp"
+#include "cgimap/backend/apidb/transaction_manager.hpp"
 
 #include <set>
