+++ include/cgimap/backend/apidb/changeset_upload/way_updater.hpp
@@ -5,7 +5,7 @@
 
 #include "cgimap/api06/changeset_upload/osmchange_tracking.hpp"
 #include "cgimap/api06/changeset_upload/way_updater.hpp"
-#include "cgimap/backend/apidb/changeset_upload/transaction_manager.hpp"
+#include "cgimap/backend/apidb/transaction_manager.hpp"
 #include "types.hpp"
 #include "util.hpp"
