+++ src/Makefile.am
@@ -197,11 +197,11 @@ libcgimap_apidb_la_SOURCES=\
 	backend/apidb/changeset.cpp \
 	backend/apidb/quad_tile.cpp \
 	backend/apidb/oauth_store.cpp \
+	backend/apidb/transaction_manager.cpp \
 	backend/apidb/utils.cpp \
 	backend/apidb/changeset_upload/changeset_updater.cpp \
 	backend/apidb/changeset_upload/node_updater.cpp \
 	backend/apidb/changeset_upload/relation_updater.cpp \
-	backend/apidb/changeset_upload/transaction_manager.cpp \
 	backend/apidb/changeset_upload/way_updater.cpp
 
 libcgimap_apidb_la_LIBADD=libcgimap_core.la @BOOST_DATE_TIME_LIB@ @LIBPQXX_LIBS@
